Extract all FluentAssertions specialized extensions to own project
