+++ Obfuscar/Properties/AssemblyInfo.cs
@@ -29,6 +29,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 [assembly: InternalsVisibleTo("ObfuscarTest")]
+[assembly: InternalsVisibleTo("ObfuscarTest.FluentAssertions")]
 
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
