+++ ObfuscarTest.FluentAssertions/Mono.Cecil/EventDefinitionAssertions.cs
@@ -9,19 +9,18 @@ using Obfuscar;
 namespace ObfuscarTest.FluentAssertions.Mono.Cecil
 {
     /// <summary>
-    /// <see cref="TypeDefinition"/> assertions based on <inheritdoc cref="ReferenceTypeAssertions{TSubject,TAssertions}"/>
+    /// <see cref="EventDefinition"/> assertions based on <inheritdoc cref="ReferenceTypeAssertions{TSubject,TAssertions}"/>
     /// </summary>
-    public class MethodDefinitionAssertions : ReferenceTypeAssertions<MethodDefinition, MethodDefinitionAssertions>
+    public class EventDefinitionAssertions : ReferenceTypeAssertions<EventDefinition, EventDefinitionAssertions>
     {
         /// <summary> Constructor </summary>
         /// <param name="instance">The instance to assert.</param>
-        public MethodDefinitionAssertions(MethodDefinition instance)
+        public EventDefinitionAssertions(EventDefinition instance)
             : base(instance)
         { }
 
         /// <inheritdoc cref = "ReferenceTypeAssertions{TSubject,TAssertions}.Identifier" />
-        protected override string Identifier => "MethodDefinition";
-
+        protected override string Identifier => "PropertyDefinition";
 
         /// <summary>
         /// Assert the <see cref="ReferenceTypeAssertions{TSubject,TAssertions}.Subject"/> to have a specific obfuscation status
@@ -35,23 +34,23 @@ namespace ObfuscarTest.FluentAssertions.Mono.Cecil
         /// </param>
         /// <param name="map">The obfuscation map to use for lookup.</param>
         /// <param name="status">The obfuscation status to compare with.</param>
-        internal AndConstraint<MethodDefinitionAssertions> HaveObfuscationStatus(
-            ObfuscationMap map, ObfuscationStatus status,
+        internal AndConstraint<EventDefinitionAssertions> HaveObfuscationStatus(
+            ObfuscationMap map, ObfuscationStatus status, 
             string because = "", params object[] becauseArgs)
         {
             Execute.Assertion.AddReportable("status",
                 () => Subject != null
-                    ? map.GetMethod(new MethodKey(new TypeKey(Subject.DeclaringType), Subject))?.Status.ToString()
+                    ? map.GetEvent(new EventKey(Subject))?.Status.ToString()
                     : null);
 
             Execute.Assertion
                 .BecauseOf(because, becauseArgs)
-                .ForCondition(map.GetMethod(new MethodKey(new TypeKey(Subject.DeclaringType), Subject))?.Status == status)
+                .ForCondition(map.GetEvent(new EventKey(Subject))?.Status == status)
                 .FailWith("Expected {context:" + Identifier + "} " +
-                          "for method {0} to have obfuscation status {1} but found status {status}.",
+                          "for event {0} to have obfuscation status {1} but found status {status}.",
                     Subject, status);
 
-            return new AndConstraint<MethodDefinitionAssertions>(this);
+            return new AndConstraint<EventDefinitionAssertions>(this);
         }
 
         /// <summary>
@@ -66,23 +65,23 @@ namespace ObfuscarTest.FluentAssertions.Mono.Cecil
         /// </param>
         /// <param name="map">The obfuscation map to use for lookup.</param>
         /// <param name="status">The obfuscation statuses to compare with.</param>
-        internal AndConstraint<MethodDefinitionAssertions> HaveObfuscationStatus(
+        internal AndConstraint<EventDefinitionAssertions> HaveObfuscationStatus(
             ObfuscationMap map, IEnumerable<ObfuscationStatus> status,
             string because = "", params object[] becauseArgs)
         {
             Execute.Assertion.AddReportable("status",
                 () => Subject != null
-                    ? map.GetMethod(new MethodKey(new TypeKey(Subject.DeclaringType), Subject))?.Status.ToString()
+                    ? map.GetEvent(new EventKey(Subject))?.Status.ToString()
                     : null);
 
             Execute.Assertion
                 .BecauseOf(because, becauseArgs)
-                .ForCondition(status.Any(s => map.GetMethod(new MethodKey(new TypeKey(Subject.DeclaringType), Subject))?.Status != s))
+                .ForCondition(status.Any(s => map.GetEvent(new EventKey(Subject))?.Status != s))
                 .FailWith("Expected {context:" + Identifier + "} " +
-                          "for method {0} to have any of obfuscation status {1} but found status {status}.",
+                          "for event {0} to have any of obfuscation status {1} but found status {status}.",
                     Subject, status);
 
-            return new AndConstraint<MethodDefinitionAssertions>(this);
+            return new AndConstraint<EventDefinitionAssertions>(this);
         }
 
         /// <summary>
@@ -97,23 +96,23 @@ namespace ObfuscarTest.FluentAssertions.Mono.Cecil
         /// </param>
         /// <param name="map">The obfuscation map to use for lookup.</param>
         /// <param name="status">The obfuscation status to compare with.</param>
-        internal AndConstraint<MethodDefinitionAssertions> NotHaveObfuscationStatus(
+        internal AndConstraint<EventDefinitionAssertions> NotHaveObfuscationStatus(
             ObfuscationMap map, ObfuscationStatus status,
             string because = "", params object[] becauseArgs)
         {
             Execute.Assertion.AddReportable("status",
                 () => Subject != null
-                    ? map.GetMethod(new MethodKey(new TypeKey(Subject.DeclaringType), Subject))?.Status.ToString()
+                    ? map.GetEvent(new EventKey(Subject))?.Status.ToString()
                     : null);
 
             Execute.Assertion
                 .BecauseOf(because, becauseArgs)
-                .ForCondition(map.GetMethod(new MethodKey(new TypeKey(Subject.DeclaringType), Subject))?.Status != status)
-                .FailWith("Expected {context:" + Identifier + "} " +
-                          "for method {0} not to have obfuscation status {1} but found status {status}.",
+                .ForCondition(map.GetEvent(new EventKey(Subject))?.Status != status)
+                .FailWith("Expected {context:" + Identifier + "} " + 
+                          "for event {0} not to have obfuscation status {1} but found status {status}.",
                     Subject, status);
 
-            return new AndConstraint<MethodDefinitionAssertions>(this);
+            return new AndConstraint<EventDefinitionAssertions>(this);
         }
 
         /// <summary>
@@ -128,23 +127,23 @@ namespace ObfuscarTest.FluentAssertions.Mono.Cecil
         /// </param>
         /// <param name="map">The obfuscation map to use for lookup.</param>
         /// <param name="status">The obfuscation statuses to compare with.</param>
-        internal AndConstraint<MethodDefinitionAssertions> NotHaveObfuscationStatus(
+        internal AndConstraint<EventDefinitionAssertions> NotHaveObfuscationStatus(
             ObfuscationMap map, IEnumerable<ObfuscationStatus> status,
             string because = "", params object[] becauseArgs)
         {
             Execute.Assertion.AddReportable("status",
                 () => Subject != null
-                    ? map.GetMethod(new MethodKey(new TypeKey(Subject.DeclaringType), Subject))?.Status.ToString()
+                    ? map.GetEvent(new EventKey(Subject))?.Status.ToString()
                     : null);
 
             Execute.Assertion
                 .BecauseOf(because, becauseArgs)
-                .ForCondition(status.All(s => map.GetMethod(new MethodKey(new TypeKey(Subject.DeclaringType), Subject))?.Status != s))
+                .ForCondition(status.All(s => map.GetEvent(new EventKey(Subject))?.Status != s))
                 .FailWith("Expected {context:" + Identifier + "} " +
-                          "for method {0} not to have and of obfuscation statuses {1} but found status {status}.",
+                          "for event {0} not to have and of obfuscation statuses {1} but found status {status}.",
                     Subject, status);
 
-            return new AndConstraint<MethodDefinitionAssertions>(this);
+            return new AndConstraint<EventDefinitionAssertions>(this);
         }
     }
 }
